Wayback paintings bot
I'm tired of all the damn links that keep breaking on paintings and are not in the Wayback Machine
Let's add a ton of links to https://archive.org/web/
The bot will take a SPARQL query that will return a bunch of links.
Bot will run over these links and if needed, add them to the Wayback machine.
This is part of https://www.wikidata.org/wiki/Wikidata:WikiProject_sum_of_all_paintings/Link_rot